[Sher] migrate MiningInfo and MinerInfo from publicData to client
